Add PostgreSQL support, import timestamps.

+++ src/main/scala/me/fornever/loglistimporter/legacy/QuoteTable.scala
@@ -2,11 +2,11 @@ package me.fornever.loglistimporter.legacy
 
 import scala.slick.driver.MySQLDriver.simple._
 
-class QuoteTable(tag: Tag) extends Table[(Int, String, Int)](tag, "list") {
+class QuoteTable(tag: Tag) extends Table[(Int, String, Long)](tag, "list") {
 
   def id = column[Int]("id", O.PrimaryKey)
   def text = column[String]("text")
-  def time = column[Int]("time")
+  def time = column[Long]("time")
 
   def * = (id, text, time)
